requirements.txt
PyYAML
eel
requests
